- Message:
Given that roles add their method to the class that composes them, it is currently impossible to have a role and a class with a submethod BUILD.

This conflicts with the entire idea of submwethods. Roles are first class entities. They should be allowed to have their own BUILD and TWEAK methods apart from the object on which they are composed.

Fortunately, in this instance, there is a workaround. Since the existing implementation has such simple logic, we simply copy it into our own role-specific constructor and call it fixed.

This should be done for EVERY role that requires GLib::Roles::Object. Consider this yet another line-wide initiative. It can be done in stages

-- Said fix now applied to ::Roles::AsyncInitable