- Workaround for GIO::Roles::File initialization issue.
Discussion:

Unfortunately this fix involvs extra code. Fortunately, that code is not THAT cumbersome, as it involves adding ancestry logic to an object that simply puns the role in question.

Still a lot of work. Will restart part 3 of the refactoring to insure that this latest code is available for all objects.

This then needs to be applied software-line wide!!