- This commit marks the end of september-refactor-review.
All changes have been made. The only dangling threads are as follows:

t/08-file.t            - Errors remain in the tests.
t/09-network-monitor.t - Errors remain in the tests.

Finishing both prior to merge could represent significant delays in the
merge, and there is no need for this. It has been decided that these
errors can be passed up to the master branch for fixes there.

This will be the first priority work for GIO post-merge